# docs/swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: メダカライブオークション - 管理者API
  description: |
    メダカライブオークションシステムの管理者向けAPI仕様書
    
    ## 認証
    すべてのエンドポイントは`Bearer`トークン認証が必要です。
    また、`admin`ロールを持つユーザーのみアクセス可能です。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: Development server
...
paths:
...
  /admin/won-items/{id}/confirm-payment:
    post:
      tags:
        - WonItems
      summary: 入金確認
      operationId: confirmPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 確認成功
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/won-items/{id}/notes:
    patch:
      tags:
        - WonItems
      summary: 備考更新
      operationId: updateNotes
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                admin_notes:
                  type: string
      responses:
        '200':
          description: 更新成功

  # ==================== Settings ====================
  /admin/settings:
    get:
      tags:
        - Settings
      summary: 設定一覧取得
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      settings:
                        type: object
    put:
      tags:
        - Settings
      summary: 設定更新
      operationId: updateSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 更新成功
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/settings/defaults:
    get:
      tags:
        - Settings
      summary: オークションデフォルト設定取得
      operationId: getAuctionDefaults
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: リソースが見つかりません。
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
...